+++ src/TwigComponent/src/DependencyInjection/TwigComponentExtension.php
@@ -121,6 +121,7 @@ final class TwigComponentExtension extends Extension implements ConfigurationInt
                 new Reference('ux.twig_component.component_renderer'),
                 new ServiceLocatorArgument(new TaggedIteratorArgument('ux.twig_component.twig_renderer', indexAttribute: 'key', needsIndexes: true)),
                 new Reference('ux.twig_component.component_stack'),
+                new Reference('twig'),
             ])
             ->addTag('twig.runtime')
         ;